Test I2C connection between Arduinos
I switched from using the Nano and Mega to using two Megas. The reason
is that the Nano clone that shipped with the transceiver uses the old
bootloader. I do not know how to select this bootloader using Arduino
CMake.

Also it is important to remember that the ports defined in
'firmware/test/CMakeLists.txt' are specific to your computer and will
have to be changed for different computers.