Change comms firmware for Arduinos to be the same
Before there was a transmitter.cpp/receiver.cpp or client.cpp/server.cpp
for the Arduinos. I think the two Arduino's can actually run the same
code and the Python scripts dictate the server/client. I am now going to
try and implement the KISS protocol in a python script.