Begin code refactor and add testing
I looked at using a library to mock the pyserial `serial` class but it
proved easier to implement a simple mock. The mocking syntax is a bit
obtuse so here is a short example about how to use it:

Other than that, I added some better error handling to the code. None of
the function are catching errors right now, and they should be modified
to do so. It is best to consult the documentation for the methods that
are called from external libraries, such as pyserial, and make sure the
appropriate errors are being caught.

HOW TO MOCK
  Mocking is done using the pytest `mocker` class. The test using
  `mocker` must receive it in the function handle like:
    `def test_setupSerial_serialPortSucceeds(mocker):`
  Let's say that we wanted to mock the `serial.Serial` function because
  that is what is being called in `serialSetup`. You would do so
  by passing `serial` to `mocker` as:
    mocker.patch.object(serial, 'Serial')
  which specifies to mock the `Serial` method in `serial`. Then you can
  pass `serial` to setupSerial and check that it was called with the
  appropriate parameters which is done with
  `serial.Serial.assert_called_with`.

  You will also see that I used the `side_effect` parameter with the
  mocked `serial.Serial`. This was to be able to return an exception.
  For returning regular value such as string or a number you would use
  `return_value`. This would be the case with, for example:
    `serial.read.return_value = "helloworld\n\r"
  Whatever calls `serial.read()` will then receive "helloworld\n\r"